Issue #377 - use the VMUSB drivers in place rather than as copies
Import support for the V1495 logic module with scaler firmware
from vmusb/daqconfig rather than from unmaintainable local copies.
