+++ main/mvlc/Makefile.am
@@ -76,11 +76,11 @@ IMPORTED_MODULES=$(IMPORT_DIR)/CMQDC32RdoHdwr.h \
 	$(IMPORT_DIR)/CNADC2530.h $(IMPORT_DIR)/CNADC2530.cpp \
 	$(IMPORT_DIR)/CCAENV1x90Opcodes.h $(IMPORT_DIR)/CCAENV1x90Registers.h \
 	$(IMPORT_DIR)/CV1x90.cpp $(IMPORT_DIR)/CV1x90.h \
-	$(IMPORT_DIR)/ CV977.h $(IMPORT_DIR)/CV977.cpp 
+	$(IMPORT_DIR)/ CV977.h $(IMPORT_DIR)/CV977.cpp \
+	$(IMPORT_DIR)/CV1495sc.h $(IMPORT_DIR)/CV1495sc.cpp 
 
 
 LOCAL_MODULES= \
-	CV1495sc.h CV1495sc.cpp \
 	CV1729.h CV1729.cpp \
 	CXLM.h CXLM.cpp CXLMTimestamp.cpp CXLMTimestamp.h \
 	CXLMFERA.cpp CXLMFERA.h \
